+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # v1.0.1
 
 - [fix] Send message using enter button.
+- [refactor] Server.
+- [refactor] Add dot env.
 
 # v1.0.0